# api/openapi.yaml
openapi: 3.0.0
info:
  title: SwiftRemit API Service
  version: 1.0.0
  description: API service for SwiftRemit currency configuration and anchor management
  contact:
    name: SwiftRemit Team
servers:
  - url: http://localhost:3000
    description: Development server
  - url: https://api.swiftremit.com
    description: Production server
...
paths:
...
  /api/anchors:
    get:
      tags:
        - Anchors
      summary: Get all anchor providers
      description: Returns all available anchor providers with optional filtering
      operationId: getAnchors
      parameters:
        - name: status
          in: query
          description: Filter by anchor status
          schema:
            type: string
            enum: [active, inactive, maintenance]
        - name: currency
          in: query
          description: Filter by supported currency
          schema:
            type: string
            maxLength: 12
      responses:
        '200':
          description: List of anchor providers
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnchorListResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/anchors/{id}:
    get:
      tags:
        - Anchors
      summary: Get anchor by ID
      description: Returns a specific anchor provider by ID
      operationId: getAnchorById
      parameters:
        - name: id
          in: path
          required: true
          description: Anchor ID
          schema:
            type: string
      responses:
        '200':
          description: Anchor details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnchorDetailResponse'
        '404':
          description: Anchor not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/anchors/admin:
    post:
      tags:
        - Anchors
      summary: Create new anchor (Admin)
      description: Creates a new anchor provider. Requires admin API key.
      operationId: createAnchor
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnchorProvider'
      responses:
        '201':
          description: Anchor created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnchorDetailResponse'
        '400':
          description: Invalid anchor payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Admin API key for protected endpoints
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - timestamp
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - message
          properties:
            message:
              type: string
            code:
              type: string
        timestamp:
          type: string
          format: date-time
    AnchorFees:
      type: object
      required:
        - deposit_fee_percent
        - withdrawal_fee_percent
      properties:
        deposit_fee_percent:
          type: number
          format: float
        deposit_fee_fixed:
          type: number
          format: float
        withdrawal_fee_percent:
          type: number
          format: float
        withdrawal_fee_fixed:
          type: number
          format: float
        min_fee:
          type: number
          format: float
        max_fee:
          type: number
          format: float
    AnchorLimits:
      type: object
      required:
        - min_amount
        - max_amount
      properties:
        min_amount:
          type: number
          format: float
        max_amount:
          type: number
          format: float
        daily_limit:
          type: number
          format: float
        monthly_limit:
          type: number
          format: float
    AnchorCompliance:
      type: object
      required:
        - kyc_required
        - kyc_level
        - supported_countries
        - restricted_countries
        - documents_required
      properties:
        kyc_required:
          type: boolean
        kyc_level:
          type: string
          enum: [basic, intermediate, advanced]
        supported_countries:
          type: array
          items:
            type: string
        restricted_countries:
          type: array
          items:
            type: string
        documents_required:
          type: array
          items:
            type: string
    AnchorProvider:
      type: object
      required:
        - id
        - name
        - domain
        - description
        - status
        - fees
        - limits
        - compliance
        - supported_currencies
        - processing_time
        - verified
      properties:
        id:
          type: string
          example: anchor-1
        name:
          type: string
          example: Example Anchor
        domain:
          type: string
          example: anchor.example.com
        logo_url:
          type: string
          format: uri
        description:
          type: string
        status:
          type: string
          enum: [active, inactive, maintenance]
        fees:
          $ref: '#/components/schemas/AnchorFees'
        limits:
          $ref: '#/components/schemas/AnchorLimits'
        compliance:
          $ref: '#/components/schemas/AnchorCompliance'
        supported_currencies:
          type: array
          items:
            type: string
        processing_time:
          type: string
          example: 1-2 business days
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
        total_transactions:
          type: integer
        verified:
          type: boolean
    AnchorListResponse:
      type: object
      required:
        - success
        - data
        - count
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        data:
          type: array
          items:
            $ref: '#/components/schemas/AnchorProvider'
        count:
          type: integer
        timestamp:
          type: string
          format: date-time
    AnchorDetailResponse:
      type: object
      required:
        - success
        - data
        - timestamp
      properties:
        success:
          type: boolean
          example: true
        data:
          $ref: '#/components/schemas/AnchorProvider'
        timestamp:
          type: string
          format: date-time
...